update/WK-7 - Update entities
