+++ src/main/java/com/github/blog/repository/entity/Post.java
@@ -1,4 +1,4 @@
-package com.github.blog.model;
+package com.github.blog.repository.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,65 +8,57 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.PrePersist;
-import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
 
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "\"user\"", schema = "blog")
-public class User {
+@Table(name = "post", schema = "blog")
+public class Post {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false)
     private Long id;
 
-    @Column(nullable = false, length = Integer.MAX_VALUE)
-    private String username;
+    @ManyToOne(optional = false)
+    @OnDelete(action = OnDeleteAction.SET_NULL)
+    @JoinColumn(name = "user_id")
+    private User user;
 
     @Column(nullable = false, length = Integer.MAX_VALUE)
-    private String password;
+    private String title;
 
     @Column(nullable = false, length = Integer.MAX_VALUE)
-    private String email;
+    private String content;
 
     @Column(nullable = false)
     private OffsetDateTime createdAt;
 
-    @Column
-    private OffsetDateTime updatedAt;
+    @OneToOne(mappedBy = "post")
+    private Order order;
 
     @ManyToMany
-    @JoinTable(name = "user_role", schema = "blog", joinColumns = {@JoinColumn(name = "user_id")},
-            inverseJoinColumns = {@JoinColumn(name = "role_id")})
-    private Set<Role> roles = new HashSet<>();
+    @JoinTable(name = "post_tag", schema = "blog", joinColumns = {@JoinColumn(name = "post_id")}, inverseJoinColumns = {@JoinColumn(name = "tag_id")})
+    private List<Tag> tags = new ArrayList<>();
 
-    @OneToOne(mappedBy = "user")
-    private UserInfo userInfo;
-
-    @OneToMany(mappedBy = "user")
-    private List<Order> orders = new ArrayList<>();
+    @OneToMany(mappedBy = "post", orphanRemoval = true)
+    private List<Comment> comments = new ArrayList<>();
 
     @PrePersist
     private void prePersist() {
         createdAt = OffsetDateTime.now();
-        updatedAt = OffsetDateTime.now();
-    }
-
-    @PreUpdate
-    private void preUpdate() {
-        updatedAt = OffsetDateTime.now();
     }
 }
