+++ src/main/java/com/github/blog/repository/entity/Reaction.java
@@ -1,23 +1,23 @@
-package com.github.blog.model;
+package com.github.blog.repository.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToMany;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.HashSet;
-import java.util.Set;
-
+/**
+ * @author Raman Haurylau
+ */
 @Getter
 @Setter
 @Entity
-@Table(name = "role", schema = "blog")
-public class Role {
+@Table(name = "reaction", schema = "blog")
+public class Reaction {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -27,6 +27,9 @@ public class Role {
     @Column(nullable = false, length = Integer.MAX_VALUE)
     private String name;
 
-    @ManyToMany(mappedBy = "roles")
-    private Set<User> users = new HashSet<>();
+    @OneToOne(mappedBy = "reaction")
+    private CommentReaction commentReaction;
+
+    @OneToOne(mappedBy = "reaction")
+    private PostReaction postReactionReaction;
 }
