+++ src/main/java/com/github/blog/repository/entity/User.java
@@ -1,4 +1,4 @@
-package com.github.blog.model;
+package com.github.blog.repository.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,57 +8,65 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.PrePersist;
+import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
 
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "post", schema = "blog")
-public class Post {
+@Table(name = "\"user\"", schema = "blog")
+public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false)
     private Long id;
 
-    @ManyToOne(optional = false)
-    @OnDelete(action = OnDeleteAction.SET_NULL)
-    @JoinColumn(name = "user_id")
-    private User user;
+    @Column(nullable = false, length = Integer.MAX_VALUE)
+    private String username;
 
     @Column(nullable = false, length = Integer.MAX_VALUE)
-    private String title;
+    private String password;
 
     @Column(nullable = false, length = Integer.MAX_VALUE)
-    private String content;
+    private String email;
 
     @Column(nullable = false)
     private OffsetDateTime createdAt;
 
-    @OneToOne(mappedBy = "post")
-    private Order order;
+    @Column
+    private OffsetDateTime updatedAt;
 
     @ManyToMany
-    @JoinTable(name = "post_tag", schema = "blog", joinColumns = {@JoinColumn(name = "post_id")}, inverseJoinColumns = {@JoinColumn(name = "tag_id")})
-    private List<Tag> tags = new ArrayList<>();
+    @JoinTable(name = "user_role", schema = "blog", joinColumns = {@JoinColumn(name = "user_id")},
+            inverseJoinColumns = {@JoinColumn(name = "role_id")})
+    private Set<Role> roles = new HashSet<>();
 
-    @OneToMany(mappedBy = "post", orphanRemoval = true)
-    private List<Comment> comments = new ArrayList<>();
+    @OneToOne(mappedBy = "user")
+    private UserInfo userInfo;
+
+    @OneToMany(mappedBy = "user")
+    private List<Order> orders = new ArrayList<>();
 
     @PrePersist
     private void prePersist() {
         createdAt = OffsetDateTime.now();
+        updatedAt = OffsetDateTime.now();
+    }
+
+    @PreUpdate
+    private void preUpdate() {
+        updatedAt = OffsetDateTime.now();
     }
 }
