+++ src/main/java/com/github/blog/repository/entity/util/OrderState.java
@@ -1,0 +1,8 @@
+package com.github.blog.repository.entity.util;
+
+/**
+ * @author Raman Haurylau
+ */
+public enum OrderState {
+    NEW, COMMITED, CANCELED, PLACED
+}
